pyproject.toml
[tool.poetry]
name = "uparking"
version = "0.1.0"
description = ""
authors = ["Christopher Sepúlveda <christopheralexander.sepulveda@alumnos.ulagos.cl>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg = "^3.1.19"
python-dotenv = "^1.0.1"
psycopg-binary = "^3.1.19"
dj-rest-auth = {extras = ["with-social"], version = "^6.0.0"}
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = {extras = ["sidecar"], version = "^0.27.2"}
python-lsp-isort = "^0.1"
djangorestframework-gis = "^1.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
pylsp-mypy = "^0.6.8"
python-lsp-server = {extras = ["all"], version = "^1.11.0"}
python-lsp-black = "^2.0.0"
django-stubs = {extras = ["compatible-mypy"], version = "^5.0.0"}
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.15.0"}
pylsp-rope = "^0.1.16"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[tool.django-stubs]
django_settings_module = "uparking.settings"
